# spec/openapi.yaml
openapi: 3.0.1
info:
  description: An API suite providing read only access to search and retrieve
    public company data
  title: Companies House Public Data API
  version: 1.0.0
servers:
  - url: https://api.company-information.service.gov.uk/
  - url: http://api.company-information.service.gov.uk/
security:
  - api_key: []
...
paths:
...
  "/company/{company_number}/filing-history/{transaction_id}":
    get:
      description: Get the filing history item of a company
      operationId: companyCompanyNumberFilingHistoryTransactionIdGET
      parameters:
        - description: The company number that the single filing is required for.
          in: path
          name: company_number
          required: true
          schema:
            type: string
        - description: The transaction id that the filing history is required for.
          in: path
          name: transaction_id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  action_date:
                    type: string
                  barcode:
                    type: string
                  category:
                    type: string
                  date:
                    type: string
                  description:
                    type: string
                  description_values:
                    type: object
                    properties:
                      made_up_date:
                        type: string
                      appointment_date:
                        type: string
                      officer_name:
                        type: string
                      psc_name:
                        type: string
                      notification_date:
                        type: string
                      old_address:
                        type: string
                      new_address:
                        type: string
                      change_date:
                        type: string
                      new_date:
                        type: string
                      termination_date:
                        type: string
                      description:
                        type: string
                      cessation_date:
                        type: string
                      charge_number:
                        type: string
                      charge_creation_date:
                        type: string
                      change_address:
                        type: string
                      date:
                        type: string
                      capital:
                        type: array
                        items:
                          type: object
                          properties:
                            currency:
                              type: string
                            figure:
                              type: string
                          required:
                            - currency
                            - figure
                      withdrawal_date:
                        type: string
                      change_details:
                        type: string
                      branch_number:
                        type: string
                      change_type:
                        type: string
                      representative_details:
                        type: string
                      brought_down_date:
                        type: string
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                      document_metadata:
                        type: string
                    required:
                      - self
                  pages:
                    type: integer
                  transaction_id:
                    type: string
                  type:
                    type: string
                  subcategory:
                    type: string
                  paper_filed:
                    type: boolean
                  associated_filings:
                    type: array
                    items:
                      type: object
                      properties:
                        category:
                          type: string
                        date:
                          type: string
                        description:
                          type: string
                        type:
                          type: string
                        action_date:
                          type: integer
                        description_values:
                          type: object
                          properties:
                            capital:
                              type: array
                              items:
                                type: object
                                properties:
                                  currency:
                                    type: string
                                  figure:
                                    type: string
                                required:
                                  - currency
                                  - figure
                            date:
                              type: string
                            capital_currency:
                              type: string
                            capital_figure:
                              type: string
                            resolution_date:
                              type: string
                        original_description:
                          type: string
                        data:
                          type: object
                          properties:
                            description:
                              type: string
                            description_values:
                              type: object
                              properties:
                                description:
                                  type: string
                              required:
                                - description
                        subcategory:
                          type: string
                        matched-default:
                          type: string
                      required:
                        - date
                        - description
                        - type
                  resolutions:
                    type: array
                    items:
                      type: object
                      properties:
                        category:
                          type: string
                        delta_at:
                          type: string
                        description:
                          type: string
                        description_values:
                          type: object
                          properties:
                            resolution_date:
                              type: string
                            description:
                              type: string
                            res_type:
                              type: string
                            case_start_date:
                              type: string
                        subcategory:
                          anyOf:
                            - type: string
                            - type: array
                              items:
                                type: string
                        type:
                          type: string
                        barcode:
                          type: string
                        original_description:
                          type: string
                      required:
                        - category
                        - description
                        - subcategory
                        - type
                  annotations:
                    type: array
                    items:
                      type: object
                      properties:
                        annotation:
                          type: string
                        category:
                          type: string
                        date:
                          type: string
                        description:
                          type: string
                        description_values:
                          type: object
                          properties:
                            description:
                              type: string
                          required:
                            - description
                        type:
                          type: string
                      required:
                        - annotation
                        - category
                        - date
                        - description
                        - description_values
                        - type
                required:
                  - category
                  - date
                  - description
                  - links
                  - transaction_id
                  - type
                additionalProperties: false
                title: filingHistoryItem
                example:
                  action_date: 2021-08-28
                  barcode: XB4S9OZF
                  category: accounts
                  date: 2022-05-25
                  description: accounts-with-accounts-type-total-exemption-full
                  description_values:
                    made_up_date: 2021-08-28
                  links:
                    self: /company/10643866/filing-history/MzM0MDQzNzkxNWFkaXF6a2N4
                    document_metadata: https://frontend-doc-api.company-information.service.gov.uk/document/c35R8ARtP6onZCeLvJwFNBTPZOclMs4c5HK2-I-VZD4
                  pages: 7
                  transaction_id: MzM0MDQzNzkxNWFkaXF6a2N4
                  type: AA
          description: Filing history items resource returned
        "401":
          content:
            application/json:
              schema:
                title: error
                type: object
          description: Unauthorised
        "404":
          content: {}
          description: Filing history not available for this company
      summary: filingHistoryItem resource
      tags:
        - filingHistory
      x-eov-operation-handler: controllers/FilingHistoryController
...
  "/company/{company_number}/exemptions":
    get:
      description: Company exemptions information
      operationId: companyCompanyNumberExemptionsGET
      parameters:
        - description: The company number that the exemptions list is quired for.
          in: path
          name: company_number
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  etag:
                    type: string
                  exemptions:
                    type: object
                    properties:
                      psc_exempt_as_trading_on_regulated_market:
                        type: object
                        properties:
                          exemption_type:
                            type: string
                          items:
                            type: array
                            items:
                              type: object
                              properties:
                                exempt_to:
                                  type: string
                                exempt_from:
                                  type: string
                              required:
                                - exempt_to
                                - exempt_from
                        required:
                          - exemption_type
                          - items
                      disclosure_transparency_rules_chapter_five_applies:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              type: object
                              properties:
                                exempt_from:
                                  type: string
                                exempt_to:
                                  type: string
                              required:
                                - exempt_from
                                - exempt_to
                          exemption_type:
                            type: string
                        required:
                          - items
                          - exemption_type
                      psc_exempt_as_trading_on_uk_regulated_market:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              type: object
                              properties:
                                exempt_from:
                                  type: string
                                exempt_to:
                                  type: string
                              required:
                                - exempt_from
                          exemption_type:
                            type: string
                        required:
                          - items
                          - exemption_type
                      psc_exempt_as_trading_on_eu_regulated_market:
                        type: object
                        properties:
                          exemption_type:
                            type: string
                          items:
                            type: array
                            items:
                              type: object
                              properties:
                                exempt_from:
                                  type: string
                                exempt_to:
                                  type: string
                              required:
                                - exempt_from
                                - exempt_to
                        required:
                          - exemption_type
                          - items
                  kind:
                    type: string
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                    required:
                      - self
                required:
                  - etag
                  - exemptions
                  - kind
                  - links
                additionalProperties: false
                title: getExemptions
                example:
                  etag: d427c6d903475de1b7d15e352ece0b1e0ac99554
                  exemptions:
                    psc_exempt_as_trading_on_regulated_market:
                      exemption_type: psc-exempt-as-trading-on-regulated-market
                      items:
                        - exempt_to: 2021-08-13
                          exempt_from: 2017-08-13
                    disclosure_transparency_rules_chapter_five_applies:
                      items:
                        - exempt_from: 2016-08-13
                          exempt_to: 2021-08-24
                      exemption_type: disclosure-transparency-rules-chapter-five-applies
                    psc_exempt_as_trading_on_uk_regulated_market:
                      items:
                        - exempt_from: 2021-08-13
                      exemption_type: psc-exempt-as-trading-on-uk-regulated-market
                  kind: exemptions
                  links:
                    self: /company/00084492/exemptions
          description: Successful response
        "401":
          content: {}
          description: Unauthorised
        "404":
          content: {}
          description: Resource not found
      tags:
        - exemptions
      x-eov-operation-handler: controllers/ExemptionsController
...
components:
...
  securitySchemes:
    api_key:
      in: header
      name: api_key
      type: apiKey
...